requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
alembic
pydantic
asyncpg
pyjwt[crypto]
bcrypt
python-multipart
ruff
pytest
PyYAML
apscheduler
aiokafka